client_side/requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.2
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
firebase-admin==7.0.0  # Firebase Admin SDK for token verification

# AWS Services
boto3==1.35.21
botocore==1.35.21

# HTTP Client
httpx==0.27.2

# Utilities
python-dateutil==2.9.0
uuid==1.30

